% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rslc}
\alias{rslc}
\title{Robust Single-Linkage Clustering}
\usage{
rslc(dists, num_clusters, min_size)
}
\arguments{
\item{dists}{distance matrix (square and symmetric)}

\item{num_clusters}{the number of clusters to find}

\item{min_size}{the minimum size of a cluster}
}
\value{
\item{clusters}{vector with the cluster id (number) for each item}
  \item{outliers}{binary vector indicating outliers}
}
\description{
Robust Single-Linkage Clustering
}
